# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/category/": {
            "get": {
                "description": "获取下厨房全部分类",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "下厨房"
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "$ref": "#/definitions/xiachufang.Category"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/category/{no}": {
            "get": {
                "description": "获取下厨房某个分类菜谱",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "下厨房"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "分类",
                        "name": "no",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "排序方式 默认:最近流行 pop:最受欢迎 time:评分",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "获取第几页, 默认第一页",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "$ref": "#/definitions/xiachufang.Category"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/recipe/{no}": {
            "get": {
                "description": "下厨房 获取菜谱",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "下厨房"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "菜谱编号",
                        "name": "no",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/xiachufang.Recipe"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/search/{keyword}": {
            "get": {
                "description": "下厨房搜索 可搜索分类/菜谱..",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "下厨房"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键字",
                        "name": "keyword",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "获取第几页, 默认第一页",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/xiachufang.RecipeListResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "xiachufang.Author": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "xiachufang.Category": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/xiachufang.Category"
                    }
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "no": {
                    "type": "string"
                }
            }
        },
        "xiachufang.Recipe": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/xiachufang.Author"
                    }
                },
                "category": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/xiachufang.Category"
                    }
                },
                "cooked": {
                    "description": "正常搜索菜谱展示",
                    "type": "integer"
                },
                "cooked_last_7_day": {
                    "description": "分类搜索时展示",
                    "type": "integer"
                },
                "cover": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "materials": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/xiachufang.RecipeMaterial"
                    }
                },
                "name": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/xiachufang.RecipeStep"
                    }
                },
                "tip": {
                    "type": "string"
                }
            }
        },
        "xiachufang.RecipeListResult": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "next_page": {
                    "type": "string"
                },
                "prev_page": {
                    "type": "string"
                },
                "recipes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/xiachufang.Recipe"
                    }
                }
            }
        },
        "xiachufang.RecipeMaterial": {
            "type": "object",
            "properties": {
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "xiachufang.RecipeStep": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "step": {
                    "type": "integer"
                }
            }
        }
    }
}